Completed CRUD operations

+++ src/main/java/com/example/ToDoList/DAO/StatusRepository.java
@@ -1,8 +1,7 @@
 package com.example.ToDoList.DAO;
 
-import com.example.ToDoList.DTO.StatusDto;
 import com.example.ToDoList.entity.Status;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StatusRepository extends JpaRepository<Status, Integer> {
+public interface StatusRepository extends JpaRepository<Status, Long> {
 }
